project/requirements.txt
Django>=4.0.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
pandas>=2.1.3
openpyxl>=3.1.2